feat: bar apperance 설정 프로토콜 정의
#90
